sdm710-common: vibrator: improve the double click effect
wf_rep_count repeats the pattern twice without a break in the
middle, so the actual haptic feedback feels like a strong single
click instead of a double click.

Therefore, drop the override on wf_rep_count and use a usleep()
between two clicks to improve the effect.

While we are at it, take the sleep duration into account when
reporting the total duration to VibratorService.

Change-Id: I2878402baac548a1aa72690f4bfe96d9b7d13cdb
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>